+++ services/src/main/java/org/keycloak/broker/oidc/OAuth2IdentityProviderConfig.java
@@ -40,6 +40,7 @@ public class OAuth2IdentityProviderConfig extends IdentityProviderModel {
     public static final String CLIENT_AUTH_METHOD = "clientAuthMethod";
     public static final String CLIENT_ID = "clientId";
     public static final String CLIENT_SECRET = "clientSecret";
+    public static final String CLIENT_ASSERTION_SIGNING_ALG = "clientAssertionSigningAlg";
 
     public OAuth2IdentityProviderConfig(IdentityProviderModel model) {
         super(model);
@@ -154,11 +155,11 @@ public class OAuth2IdentityProviderConfig extends IdentityProviderModel {
     }
 
     public String getClientAssertionSigningAlg() {
-        return getConfig().get("clientAssertionSigningAlg");
+        return getConfig().get(CLIENT_ASSERTION_SIGNING_ALG);
     }
     
     public void setClientAssertionSigningAlg(String signingAlg) {
-        getConfig().put("clientAssertionSigningAlg", signingAlg);
+        getConfig().put(CLIENT_ASSERTION_SIGNING_ALG, signingAlg);
     }
 
     public boolean isPassScope() {