+++ CHANGELOG.md
@@ -16,6 +16,11 @@ Our Keycloak version is working well with PostgreSQL database. For using other S
 ### Added
 - Search in User Identity Provider Linkin
 
+### Fixed
+- Explicit Registration as RP, send authority hint from Trust chain resolution
+- Explicit Registration as RP, correct grant types sent
+- Explicit Registration as RP, correct Client authentication and Client assertion signature algorithm Identity Provider values 
+
 ## [22.0.13-1.21] - 2025-08-29
 
 ### Fixed